chore: move da classe ApiErrorArray para outro diretorio
