+++ src/main/java/fiap/tech/challenge/hospital_manager/exception/handlers/GlobalExceptionHandler.java
@@ -1,6 +1,5 @@
 package fiap.tech.challenge.hospital_manager.exception.handlers;
 
-import fiap.tech.challenge.hospital_manager.controller.exceptions.ApiErrorArray;
 import fiap.tech.challenge.hospital_manager.utils.ApiErrorBuilder;
 
 import jakarta.servlet.http.HttpServletRequest;
