+++ src/main/java/fiap/tech/challenge/hospital_manager/utils/ApiErrorBuilder.java
@@ -4,7 +4,7 @@ package fiap.tech.challenge.hospital_manager.utils;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import fiap.tech.challenge.hospital_manager.controller.exceptions.ApiErrorArray;
+import fiap.tech.challenge.hospital_manager.exception.handlers.ApiErrorArray;
 
 import java.time.LocalDateTime;
 import java.util.List;
